+++ src/Type/CommunityChatRemoved.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Phptg\BotApi\Type;
+
+/**
+ * @see https://core.telegram.org/bots/api#communitychatremoved
+ *
+ * @api
+ */
+final readonly class CommunityChatRemoved {}